requirements.txt
flask==3.0.3
flask-pymongo==2.3.0
pymongo==4.7.2
python-dotenv==1.0.1
marshmallow==3.21.3
celery==5.4.0
redis==5.0.7
gunicorn==22.0.0
pytest==8.2.2
pytest-flask==1.3.0
mongomock==4.1.2
setuptools>=69.0
